set DemoCards build dir to "docs/src" instead of "docs/src/democards"
For safety consideration, the first version of DemoCards set intermediate build dir
to be `docs/src/democards`. This is pretty good and safe, the only drawback of it is
that the generated URL contains a "democards" in it. For example,

"https://juliaimages.org/latest/democards/examples/"

This commit removes this and sets it back to

"https://juliaimages.org/latest/examples/"

as if everything is just there in "docs/src" (but it is not)

Note 1:

This can be potentially dangerous because there might be name conflicts, I choose to remove
the old existing folders, and if the user pass a wrong "root" keyword, removing that folder
could possibly remove some important data.

A user prompt is trigger when such case happens. If things goes smoothly, that prompt should
never be trigger.

Note 2:

This also means that user can not put their democards contents in "docs/src" folder, an error
would be thrown in such cases. For example:

ERROR: ArgumentError: Demo page /Users/jc/Documents/Julia/DemoCards/docs/src/quickstart should not be placed at "docs/src" folder.