Use EditURL instead of regex for edit links
Previously Documenter's HTML was postprocessed in order to point edit on
GitHub links to the original source file as input to DemoCards.jl. This
commit changes this so that the original source file is provided to
Documenter.jl in the generated markdown files using the @meta: EditURL
setting. This is less fragile to changes in Documenter.jl or e.g.
different templates since it doesn't rely on using regex to modify the
particular form of the HTML output by Documenter.jl.